refactor(client-docs): split components and logic for improved maintainability
- Moved modal logic into a separate component.
- Extracted API calls and utility functions for better organization.
- Simplified table rendering and event handlers.
- Enhanced readability and prepared for adding future features.
