+++ src/renderer/constants/constantData.tsx
@@ -1,0 +1,14 @@
+export const articles = [
+    { label: 'Michelin Tire', value: 'Michelin Tire' },
+    { label: 'Bridgestone Tire', value: 'Bridgestone Tire' },
+    { label: 'Goodyear Tire', value: 'Goodyear Tire' },
+    { label: 'Car Accessory A', value: 'Car Accessory A' },
+    { label: 'Car Accessory B', value: 'Car Accessory B' },
+  ];
+  
+  export const carRepairServices = [
+    { label: 'Oil Change', value: 'Oil Change' },
+    { label: 'Tire Rotation', value: 'Tire Rotation' },
+    { label: 'Engine Tuning', value: 'Engine Tuning' },
+  ];
+
